Move kernel dictionary infrastructure to `kernel` package
