+++ src/eradiate/kernel/__init__.pyi
@@ -3,3 +3,9 @@ from . import logging as logging
 from . import transform as transform
 from ._bitmap import bitmap_to_dataset as bitmap_to_dataset
 from ._bsdf import eval_bsdf as eval_bsdf
+from ._kernel_dict import InitParameter as InitParameter
+from ._kernel_dict import KernelDictTemplate as KernelDictTemplate
+from ._kernel_dict import Parameter as Parameter
+from ._kernel_dict import ParamFlags as ParamFlags
+from ._kernel_dict import UpdateMapTemplate as UpdateMapTemplate
+from ._kernel_dict import UpdateParameter as UpdateParameter
