+++ src/eradiate/scenes/core.py
@@ -1,14 +1,11 @@
 from __future__ import annotations
 
-import enum
 import importlib
 import typing as t
 from abc import ABC, abstractmethod
-from collections import UserDict
 from typing import Mapping, Sequence
 
 import attrs
-import mitsuba as mi
 import numpy as np
 import pint
 import pinttr
@@ -16,350 +13,15 @@ from pinttr.util import ensure_units
 
 from .._factory import Factory
 from ..attrs import documented, parse_docs
-from ..contexts import KernelDictContext
 from ..exceptions import TraversalError
+from ..kernel import (
+    KernelDictTemplate,
+    Parameter,
+    ParameterMap,
+    UpdateMapTemplate,
+)
 from ..units import unit_context_config as ucc
 from ..units import unit_registry as ureg
-from ..util.misc import nest
-
-# ------------------------------------------------------------------------------
-#                             Kernel dictionary
-# ------------------------------------------------------------------------------
-
-
-class ParamFlags(enum.Flag):
-    """
-    Parameter flags.
-    """
-
-    NONE = 0
-    INIT = enum.auto()  #: Used for kernel dict generation
-    SPECTRAL = enum.auto()  #: Varies during the spectral loop
-    GEOMETRIC = enum.auto()  #: Triggers a scene rebuild
-    ALL = INIT | SPECTRAL | GEOMETRIC
-    UPDATE = SPECTRAL | GEOMETRIC  #: Update-specific flags
-
-
-class _Unused:
-    """
-    Sentinel class to indicate when a parameter is unused, accessible as
-    :data:`.Param.UNUSED`.
-
-    Notes
-    -----
-    ``bool(_Unused)`` evaluates to ``False``.
-    """
-
-    _singleton = None
-
-    def __new__(cls):
-        if _Unused._singleton is None:
-            _Unused._singleton = super(_Unused, cls).__new__(cls)
-        return _Unused._singleton
-
-    def __repr__(self):
-        return "UNUSED"
-
-    def __bool__(self):
-        return False
-
-    def __len__(self):
-        return 0
-
-
-@parse_docs
-@attrs.define
-class Parameter:
-    """
-    This class declares an Eradiate parameter. It holds a evaluation protocol
-    for a Mitsuba scene parameter and exposes an interface to generate update
-    values during the execution of a parametric loop.
-    """
-
-    #: Sentinel value indicated that a parameter is not used
-    UNUSED: t.ClassVar = _Unused()
-
-    evaluator: t.Callable = documented(
-        attrs.field(repr=False),
-        doc="A callable that returns the value of the parameter for a given "
-        "context, with signature ``f(ctx: KernelDictContext) -> Any``.",
-        type="callable",
-    )
-
-    flags: ParamFlags = documented(
-        attrs.field(default=ParamFlags.ALL),
-        doc="Flags specifying parameter attributes. By default, the declared "
-        "parameter will pass all filters.",
-        type=".ParaFlags",
-        default=".ParamFlags.ALL",
-    )
-
-    lookup_id: t.Optional[t.Callable[[mi.Object, str], t.Optional[str]]] = documented(
-        attrs.field(default=None),
-        doc="A callable that searches a Mitsuba scene tree node for a desired "
-        "parameter ID.",
-        type="callable or None",
-        init_type="callable, optional",
-        default="None",
-    )
-
-    parameter_id: t.Optional[str] = documented(
-        attrs.field(default=None),
-        doc="The full ID of the Mitsuba scene parameter to update.",
-        type="str or None",
-        init_type="str, optional",
-    )
-
-    def __call__(self, ctx: KernelDictContext) -> t.Any:
-        return self.evaluator(ctx)
-
-
-@attrs.define(slots=False)
-class ParameterMap(UserDict, ABC):
-    """
-    A dict-like structure mapping parameter paths to methods generating them.
-    """
-
-    data: dict[str, Parameter] = attrs.field(factory=dict)
-
-    def remove(self, keys: t.Union[str, t.List[str]]) -> None:
-        """
-        Remove all parameters matching the given regular expression.
-
-        Parameters
-        ----------
-        keys : str or list of str
-            Regular expressions matching the parameters to remove.
-
-        Notes
-        -----
-        This method mutates the parameter map.
-        """
-        if not isinstance(keys, list):
-            keys = [keys]
-
-        import re
-
-        regexps = [re.compile(k).match for k in keys]
-        keys = [k for k in self.keys() if any(r(k) for r in regexps)]
-
-        for key in keys:
-            del self.data[key]
-
-    def keep(self, keys: t.Union[str, t.List[str]]) -> None:
-        """
-        Keep only parameters matching the given regular expression.
-
-        Parameters
-        ----------
-        keys : str or list of str
-            Regular expressions matching the parameters to keep.
-
-        Notes
-        -----
-        This method mutates the parameter map.
-        """
-        if not isinstance(keys, list):
-            keys = [keys]
-
-        import re
-
-        regexps = [re.compile(k).match for k in keys]
-        keys = [k for k in self.keys() if any(r(k) for r in regexps)]
-        result = {k: self.data[k] for k in keys}
-        self.data = result
-
-    @abstractmethod
-    def render(
-        self,
-        ctx: KernelDictContext,
-        flags: ParamFlags = ParamFlags.ALL,
-        drop: bool = False,
-    ) -> dict:
-        """
-        Evaluate the parameter map for a set of arguments.
-
-        Parameters
-        ----------
-        ctx : :class:`.KernelDictContext`
-            A kernel dictionary context.
-
-        flags : :class:`.ParamFlags`
-            Parameter flags. Only parameters with at least one of the specified
-            will pass the filter.
-
-        drop : bool
-            If ``True``, drop unused parameters. Parameters may be unused either
-            because they were filtered out by the flags or because context
-            information implied it.
-
-        Returns
-        -------
-        dict
-
-        Raises
-        ------
-        ValueError
-            If ``drop`` is ``False`` and the rendered parameter map contains an
-            unused parameter.
-        """
-        pass
-
-
-@attrs.define
-class KernelDictTemplate(ParameterMap):
-    """
-    A dict-like structure which contains the structure of an instantiable kernel
-    dictionary.
-
-    Entries are indexed by dot-separated paths which can then be expanded to
-    a nested dictionary using the :meth:`.render` method.
-
-    Each entry can be either a hard-coded value which can be directly
-    interpreted by the :func:`mitsuba.load_dict` function, or a :class:`.Param`
-    object, which must be rendered before the template can be instantiated.
-    """
-
-    def render(
-        self,
-        ctx: KernelDictContext,
-        flags: ParamFlags = ParamFlags.INIT,
-        drop: bool = False,
-        nested: bool = True,
-    ) -> dict:
-        """
-        Render the template as a nested dictionary using a parameter map to fill
-        in empty fields.
-
-        Parameters
-        ----------
-        ctx : :class:`.KernelDictContext`
-            A kernel dictionary context.
-
-        flags : :class:`.ParamFlags`
-            Parameter flags. Only parameters with at least one of the specified
-            will pass the filter.
-
-        drop : bool
-            If ``True``, drop unused parameters. Parameters may be unused either
-            because they were filtered out by the flags or because context
-            information implied it.
-
-        nested : bool, optional
-            If ``True``, the returned dictionary will be nested and suitable for
-            instantiation by Mitsuba; otherwise, the returned dictionary will be
-            flat.
-
-        Returns
-        -------
-        dict
-        """
-        result = self.data.copy()
-        unused = render_parameters(result, ctx=ctx, flags=flags, drop=drop)
-
-        # Check for leftover empty values
-        if not drop and unused:
-            raise ValueError(f"Unevaluated parameters: {unused}")
-
-        return nest(result, sep=".") if nested else result
-
-
-@attrs.define
-class UpdateMapTemplate(ParameterMap):
-    def render(
-        self,
-        ctx: KernelDictContext,
-        flags: ParamFlags = ParamFlags.ALL,
-        drop: bool = False,
-    ) -> dict:
-        """
-        Evaluate the parameter map for a set of arguments.
-
-        Parameters
-        ----------
-        ctx : :class:`.KernelDictContext`
-            A kernel dictionary context.
-
-        flags : :class:`.ParamFlags`
-            Parameter flags. Only parameters with at least one of the specified
-            will pass the filter.
-
-        drop : bool
-            If ``True``, drop unused parameters. Parameters may be unused either
-            because they were filtered out by the flags or because context
-            information implied it.
-
-        Returns
-        -------
-        dict
-
-        Raises
-        ------
-        ValueError
-            If ``drop`` is ``False`` and the rendered parameter map contains an
-            unused parameter.
-        """
-        result, unused = render_parameters(self, ctx=ctx, flags=flags, drop=drop)
-
-        # Check for leftover empty values
-        if not drop and unused:
-            raise ValueError(f"Unevaluated parameters: {unused}")
-
-        return result
-
-
-def render_parameters(
-    pmap: t.Mapping,
-    ctx: KernelDictContext,
-    flags: ParamFlags = ParamFlags.ALL,
-    drop: bool = False,
-) -> t.Tuple[dict, list]:
-    """
-    Render parameters in a template dictionary.
-
-    Parameters
-    ----------
-    pmap : mapping
-        A dict-like containing parameters to render. *In-place* modification
-        will be performed.
-
-    ctx : .KernelDictContext
-        A kernel dictionary context.
-
-    flags : .ParamFlags
-        Parameter flags. Only parameters with at least one of the specified
-        flags will pass the filter.
-
-    drop : bool
-        If ``True``, drop unused parameters.
-
-    Returns
-    -------
-    result : dict
-        A dictionary in which parameters are evaluated.
-
-    unused : list
-        A list of unused parameters. A parameter may be unused because it was
-        filtered out by flags or because context information implied it.
-    """
-    unused = []
-    result = {}
-
-    for k in list(pmap.keys()):
-        v = pmap[k]
-
-        if isinstance(v, Parameter):
-            key = k if v.parameter_id is None else v.parameter_id
-
-            if v.flags & flags:
-                result[key] = v(ctx)
-            else:
-                unused.append(k)
-                if not drop:
-                    result[key] = Parameter.UNUSED
-
-    return result, unused
-
 
 # ------------------------------------------------------------------------------
 #                           Scene element interface
