+++ src/eradiate/scenes/measure/_core.py
@@ -14,7 +14,7 @@ import xarray as xr
 
 import eradiate
 
-from ..core import NodeSceneElement, Parameter, ParamFlags
+from ..core import NodeSceneElement
 from ..spectra import InterpolatedSpectrum, Spectrum, UniformSpectrum, spectrum_factory
 from ... import ckd, converters, validators
 from ..._factory import Factory
@@ -26,6 +26,7 @@ from ...contexts import (
     SpectralContext,
 )
 from ...exceptions import ModeError, UnsupportedModeError
+from ...kernel._kernel_dict import Parameter, ParamFlags
 from ...srf_tools import convert as convert_srf
 from ...units import PhysicalQuantity, interpret_quantities
 from ...units import unit_context_config as ucc
