+++ src/eradiate/scenes/shapes/_buffermesh.py
@@ -9,9 +9,10 @@ import pint
 import pinttr
 
 from ._core import ShapeInstance
-from ..core import Parameter, traverse
+from ..core import traverse
 from ...attrs import documented, parse_docs
 from ...contexts import KernelDictContext
+from ...kernel._kernel_dict import Parameter
 from ...units import unit_context_config as ucc
 from ...units import unit_context_kernel as uck
