+++ src/eradiate/scenes/spectra/_solar_irradiance.py
@@ -10,10 +10,10 @@ import pint
 import xarray as xr
 
 from ._core import SpectrumNode
-from ..core import Parameter, ParamFlags
 from ... import converters, data, validators
 from ...attrs import documented, parse_docs
 from ...ckd import Bindex
+from ...kernel._kernel_dict import Parameter, ParamFlags
 from ...units import PhysicalQuantity, to_quantity
 from ...units import unit_context_config as ucc
 from ...units import unit_context_kernel as uck
